requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
docker==7.0.0
idna==3.7
packaging==24.0
pycryptodome==3.20.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
urllib3==2.2.1
